// x-pack/platform/plugins/shared/ingest_pipelines/public/application/components/pipeline_editor/types.d.ts
import type { Dispatch } from 'react';
import type { OnFormUpdateArg } from '../../../shared_imports';
import type { SerializeResult } from './serialize';
import type { OnActionHandler, ProcessorInfo } from './components';
import type { ProcessorsDispatch, State as ProcessorsReducerState } from './processors_reducer';
/**
 * An array of keys that map to a value in an object
 * structure.
 *
 * For instance:
 * ['a', 'b', '0', 'c'] given { a: { b: [ { c: [] } ] } } => []
 *
 * Additionally, an empty selector `[]`, is a special indicator
 * for the root level.
 */
export type ProcessorSelector = string[];
/** @internal */
export interface ProcessorInternal<CustomProcessorOptions = {}> {
    id: string;
    type: string;
    options: {
        [key: string]: any;
    };
    onFailure?: ProcessorInternal[];
}
export type { OnFormUpdateArg };
export interface FormValidityState {
    validate: OnFormUpdateArg<any>['validate'];
}
export interface OnUpdateHandlerArg extends FormValidityState {
    getData: () => SerializeResult;
}
export type OnUpdateHandler = (arg: OnUpdateHandlerArg) => void;
/**
 * The editor can be in different modes. This enables us to hold
 * a reference to data dispatch to the reducer (like the {@link ProcessorSelector}
 * which will be used to update the in-memory processors data structure.
 */
export type EditorMode = {
    id: 'creatingProcessor';
    arg: {
        selector: ProcessorSelector;
        buttonRef?: React.RefObject<HTMLButtonElement | HTMLAnchorElement>;
    };
} | {
    id: 'movingProcessor';
    arg: ProcessorInfo;
} | {
    id: 'managingProcessor';
    arg: {
        processor: ProcessorInternal;
        selector: ProcessorSelector;
        buttonRef?: React.RefObject<HTMLButtonElement | HTMLAnchorElement>;
    };
} | {
    id: 'removingProcessor';
    arg: {
        selector: ProcessorSelector;
    };
} | {
    id: 'idle';
};
export interface ContextValueEditor {
    mode: EditorMode;
    setMode: Dispatch<EditorMode>;
}
export interface ContextValueProcessors {
    state: ProcessorsReducerState;
    dispatch: ProcessorsDispatch;
}
export interface ContextValueState {
    processors: ContextValueProcessors;
    editor: ContextValueEditor;
}
export interface ContextValue {
    onTreeAction: OnActionHandler;
    state: ContextValueState;
}
export interface Document {
    _id: string;
    [key: string]: any;
}
export type ProcessorStatus = 'success' | 'error' | 'error_ignored' | 'dropped' | 'skipped' | 'inactive';
export interface ProcessorResult {
    processor_type: string;
    status: ProcessorStatus;
    doc?: Document;
    tag: string;
    ignored_error?: any;
    error?: any;
    processorInput?: Document;
    [key: string]: any;
}
export interface ProcessorResults {
    processor_results: ProcessorResult[];
}
export interface VerboseTestOutput {
    docs: ProcessorResults[];
}