// x-pack/platform/plugins/shared/ingest_pipelines/public/application/components/pipeline_editor/components/pipeline_processors_editor_item/i18n_texts.d.ts
export declare const i18nTexts: {
    moveButtonLabelWithName: ({ processorName }: {
        processorName: string;
    }) => string;
    editButtonLabel: string;
    duplicateButtonLabel: string;
    addOnFailureButtonLabel: string;
    cancelMoveButtonLabelWithName: ({ processorName }: {
        processorName: string;
    }) => string;
    deleteButtonLabel: string;
    moreButtonAriaLabel: string;
    processorTypeLabel: ({ type }: {
        type: string;
    }) => string;
    descriptionPlaceholder: string;
    dropZoneMoveBeforeLabel: ({ movingProcessor, targetProcessor, section, }: {
        movingProcessor: string;
        targetProcessor: string;
        section: string;
    }) => string;
    dropZoneCannotMoveBeforeLabel: ({ movingProcessor, targetProcessor, section, }: {
        movingProcessor: string;
        targetProcessor: string;
        section: string;
    }) => string;
    dropZoneMoveAfterLabel: ({ movingProcessor, targetProcessor, section, }: {
        movingProcessor: string;
        targetProcessor: string;
        section: string;
    }) => string;
    dropZoneCannotMoveAfterLabel: ({ movingProcessor, targetProcessor, section, }: {
        movingProcessor: string;
        targetProcessor: string;
        section: string;
    }) => string;
};